requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
APScheduler==3.11.0
